fixing RedisCloudDatabasesResult
